move res components
