+++ libs/react/shelter-operator/src/lib/components/AppSidebar/ShelterMgmtLinks.tsx
@@ -44,6 +44,19 @@ export function ShelterMgmtLinks(props: IProps) {
         >
           Rooms
         </Sidebar.Link>
+        <Sidebar.Link
+          to={shelterMgmtRoute(
+            shelterId,
+            mgmtRouteConfig.children.reservations,
+          )}
+          isActive={isShelterMgmtRoute(pathname, {
+            segment: mgmtRouteConfig.children.reservations,
+          })}
+          collapsed={!isOpen}
+          replace
+        >
+          Reservations
+        </Sidebar.Link>
       </Sidebar.Content>
     </Sidebar.NestedLinks>
   );
