+++ libs/react/shelter-operator/src/lib/components/ShelterManagement/segments/Reservations/ReservationForm/sections/index.ts
@@ -1,0 +1,2 @@
+export { BedRoomSelection } from './BedRoomSelection';
+export { ClientSearchInput } from './ClientSearchInput';
