+++ libs/react/shelter-operator/src/lib/components/ShelterManagement/segments/Reservations/ReservationTable.tsx
@@ -1,15 +1,13 @@
+import { ReservationStatusChoices } from '@monorepo/ba-platform/types';
 import { formatClientDisplayName } from '@monorepo/react/shared';
 import { Check, X } from 'lucide-react';
 import type { CSSProperties, ReactNode } from 'react';
 import { useMemo } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { ReservationStatusChoices } from '@monorepo/ba-platform/types';
-import { ReservationsQuery } from '../hooks/useReservations/__generated__/useReservations.generated';
-import { shelterEditResourceRoute } from '../routing';
-import { Button } from './base-ui/buttons';
-import { StatusBadge } from './base-ui/status-badge/StatusBadge';
-import { Table, type TableColumn } from './base-ui/table';
-import { reservationStatusInfo } from './reservations/reservation-form/constants/reservationFormOptions';
+import { ReservationsQuery } from '../../../../hooks/useReservations/__generated__/useReservations.generated';
+import { Button } from '../../../base-ui/buttons';
+import { StatusBadge } from '../../../base-ui/status-badge/StatusBadge';
+import { Table, type TableColumn } from '../../../base-ui/table';
+import { reservationStatusInfo } from './ReservationForm/formSchema';
 
 const CONFIRM_ELIGIBLE_STATUSES: Set<ReservationStatusChoices> = new Set([
   ReservationStatusChoices.Confirmed,
@@ -22,18 +20,18 @@ const CANCEL_ELIGIBLE_STATUSES: Set<ReservationStatusChoices> = new Set([
   ReservationStatusChoices.CheckInOverdue,
 ]);
 
-type ReservationRow = NonNullable<
+export type Reservation = NonNullable<
   ReservationsQuery['reservations']['results'][number]
 >;
 
 type ReservationTableProps = {
-  reservations: ReservationRow[];
-  shelterId: string;
+  reservations: Reservation[];
   loading?: boolean;
   loadingState?: ReactNode;
   emptyState?: ReactNode;
   isConfirmActionLoading?: boolean;
   isCancelActionLoading?: boolean;
+  onEdit?: (reservationId: string) => void;
   onCheckIn?: (reservationId: string) => void;
   onComplete?: (reservationId: string) => void;
   onCancel?: (reservationId: string) => void;
@@ -50,7 +48,6 @@ type ReservationTableProps = {
 
 export function ReservationTable({
   reservations,
-  shelterId,
   loading,
   loadingState,
   emptyState = (
@@ -60,6 +57,7 @@ export function ReservationTable({
   ),
   isConfirmActionLoading,
   isCancelActionLoading,
+  onEdit,
   onCheckIn,
   onComplete,
   onCancel,
@@ -73,9 +71,7 @@ export function ReservationTable({
   rowStyle,
   trailingColumnWidth = '140px',
 }: ReservationTableProps) {
-  const navigate = useNavigate();
-
-  const columns: TableColumn<ReservationRow>[] = useMemo(
+  const columns: TableColumn<Reservation>[] = useMemo(
     () => [
       {
         key: 'client',
@@ -224,35 +220,35 @@ export function ReservationTable({
       columns={columns}
       rows={reservations}
       getRowKey={(reservation) => reservation.id}
-      getRowSlot={(rowObject) => (
+      getRowSlot={(reservation) => (
         <div
           className="flex items-center justify-end gap-1"
           onClick={(e) => e.stopPropagation()}
           role="group"
           aria-label="Reservation actions"
         >
-          {CONFIRM_ELIGIBLE_STATUSES.has(rowObject.status) && (
+          {CONFIRM_ELIGIBLE_STATUSES.has(reservation.status) && (
             <Button
               type="button"
               variant="confirm"
               className="text-[#747A82]"
               aria-label={
-                rowObject.status === ReservationStatusChoices.CheckedIn
+                reservation.status === ReservationStatusChoices.CheckedIn
                   ? 'Mark completed'
                   : 'Mark checked in'
               }
               leftIcon={<Check size={24} stroke="black" />}
               disabled={isConfirmActionLoading}
               onClick={() => {
-                if (rowObject.status === ReservationStatusChoices.CheckedIn) {
-                  onComplete?.(rowObject.id);
+                if (reservation.status === ReservationStatusChoices.CheckedIn) {
+                  onComplete?.(reservation.id);
                 } else {
-                  onCheckIn?.(rowObject.id);
+                  onCheckIn?.(reservation.id);
                 }
               }}
             />
           )}
-          {CANCEL_ELIGIBLE_STATUSES.has(rowObject.status) && (
+          {CANCEL_ELIGIBLE_STATUSES.has(reservation.status) && (
             <Button
               type="button"
               variant="trash"
@@ -260,7 +256,7 @@ export function ReservationTable({
               aria-label="Cancel reservation"
               leftIcon={<X size={24} stroke="black" />}
               disabled={isCancelActionLoading}
-              onClick={() => onCancel?.(rowObject.id)}
+              onClick={() => onCancel?.(reservation.id)}
             />
           )}
           <Button
@@ -268,9 +264,7 @@ export function ReservationTable({
             variant="edit"
             className="text-[#747A82]"
             aria-label="Edit reservation"
-            onClick={() =>
-              navigate(shelterEditResourceRoute(shelterId, 'reservation', rowObject.id))
-            }
+            onClick={() => onEdit?.(reservation.id)}
           />
         </div>
       )}
