--- libs/react/shelter-operator/src/lib/components/reservations/reservation-form/constants/defaultReservationFormData.ts
@@ -1,13 +1,0 @@
-import { ReservationStatusChoices } from '@monorepo/ba-platform/types';
-import type { ReservationFormData } from '../formTypes';
-
-/** Fresh copy of reservation form defaults to avoid shared mutable state. */
-export const createEmptyReservationFormData = (): ReservationFormData => ({
-  bedId: null,
-  roomId: null,
-  clientIds: [],
-  primaryClientId: null,
-  status: ReservationStatusChoices.Confirmed,
-  startDate: '',
-  notes: '',
-});
